+++ fe1-web/features/rollCall/network/__tests__/RollCallMessageApi.ts
@@ -1,24 +1,24 @@
 import 'jest-extended';
+import { beforeEach } from '@jest/globals';
+
 import '__tests__/utils/matchers';
-import {
-  ActionType,
-  CloseRollCall,
-  CreateRollCall,
-  MessageData,
-  ObjectType,
-  OpenRollCall,
-  ReopenRollCall,
-} from 'model/network/method/message/data';
+import { ActionType, MessageData, ObjectType } from 'model/network/method/message/data';
 import { Hash, PublicKey, Timestamp } from 'model/objects';
 import { OpenedLaoStore } from 'store';
-import { beforeEach } from '@jest/globals';
 import {
   defaultMessageDataFields, mockLao, mockLaoId,
 } from '__tests__/utils/TestUtils';
 import { publish as mockPublish } from 'network/JsonRpcApi';
-import * as msApi from '../MessageApi';
 
-jest.mock('network/JsonRpcApi');
+import * as msApi from '../RollCallMessageApi';
+import {
+  CloseRollCall,
+  CreateRollCall,
+  OpenRollCall,
+  ReopenRollCall,
+} from '../messages';
+
+jest.mock('features/rollCall/network/RollCallMessageApi.ts');
 const publishMock = mockPublish as jest.Mock;
 
 const mockEventName = 'myRollCall';
