+++ fe1-web/features/rollCall/network/messages/__tests__/CreateRollCall.test.ts
@@ -1,10 +1,12 @@
 import 'jest-extended';
+
 import '__tests__/utils/matchers';
 import { ActionType, ObjectType } from 'model/network/method/message/data/MessageData';
 import { Hash, Timestamp } from 'model/objects';
 import { mockLao, mockLaoId } from '__tests__/utils/TestUtils';
 import { ProtocolError } from 'model/network/ProtocolError';
 import { OpenedLaoStore } from 'store';
+
 import { CreateRollCall } from '../CreateRollCall';
 
 const TIMESTAMP = new Timestamp(1609455600); // 1st january 2021
