+++ fe1-web/features/rollCall/objects/index.ts
@@ -1,0 +1,1 @@
+export * from './RollCall';
