+++ fe1-web/model/network/method/message/data/index.ts
@@ -2,5 +2,4 @@ export * from './MessageData';
 export * from './Checker';
 export * from './witness';
 export * from './lao';
-export * from './rollCall';
 export * from './MessageRegistry';
