+++ fe1-web/network/MessageApi.ts
@@ -1,12 +1,8 @@
 import {
-  EventTags, Hash, Lao, PublicKey, Timestamp,
+  Hash, Lao, PublicKey, Timestamp,
 } from 'model/objects';
 import {
-  CloseRollCall,
   CreateLao,
-  CreateRollCall,
-  OpenRollCall,
-  ReopenRollCall,
   StateLao,
   UpdateLao,
   WitnessMessage,
@@ -79,106 +75,4 @@ export function requestWitnessMessage(channel: Channel, messageId: Hash): Promis
   });
 
   return publish(channel, message);
-}
-
-/**
- * Send a server query asking for the creation of a roll call
- * @param name the name of the roll call event
- * @param location the location of the roll call event
- * @param proposedStart the time at which the attendee scanning will start (pending confirmation)
- * @param proposedEnd the time at which the attendee scanning will stop (pending confirmation)
- * @param description (optional) further information regarding the event
- * @return A Promise resolving when the operation is completed
- */
-export function requestCreateRollCall(
-  name: string, location: string,
-  proposedStart: Timestamp, proposedEnd: Timestamp,
-  description?: string,
-): Promise<void> {
-  const time: Timestamp = Timestamp.EpochNow();
-  const currentLao: Lao = OpenedLaoStore.get();
-
-  const message = new CreateRollCall({
-    id: Hash.fromStringArray(
-      EventTags.ROLL_CALL, currentLao.id.toString(), time.toString(), name,
-    ),
-    name: name,
-    creation: time,
-    location: location,
-    proposed_start: Timestamp.max(time, proposedStart),
-    proposed_end: proposedEnd,
-    description: description,
-  });
-
-  const laoCh = channelFromIds(currentLao.id);
-  return publish(laoCh, message);
-}
-
-/**
- * Send a server query asking for the opening of a roll call
- * @param rollCallId the ID of the roll call
- * @param start (optional) the time at which the operation happens, defaults to now.
- * @return A Promise resolving when the operation is completed
- */
-export function requestOpenRollCall(rollCallId: Hash, start?: Timestamp): Promise<void> {
-  const lao: Lao = OpenedLaoStore.get();
-  const time = (start === undefined) ? Timestamp.EpochNow() : start;
-
-  const message = new OpenRollCall({
-    update_id: Hash.fromStringArray(
-      EventTags.ROLL_CALL, lao.id.toString(), rollCallId.toString(), time.toString(),
-    ),
-    opens: rollCallId,
-    opened_at: time,
-  });
-
-  return publish(channelFromIds(lao.id), message);
-}
-
-/**
- * Send a server query asking for the re-opening of a roll call
- * @param rollCallId the ID of the roll call
- * @param start (optional) the time at which the operation happens, defaults to now.
- * @return A Promise resolving when the operation is completed
- */
-export function requestReopenRollCall(rollCallId: Hash, start?: Timestamp): Promise<void> {
-  const lao: Lao = OpenedLaoStore.get();
-  const time = (start === undefined) ? Timestamp.EpochNow() : start;
-
-  const message = new ReopenRollCall({
-    update_id: Hash.fromStringArray(
-      EventTags.ROLL_CALL, lao.id.toString(), rollCallId.toString(), time.toString(),
-    ),
-    opens: rollCallId,
-    opened_at: time,
-  });
-
-  return publish(channelFromIds(lao.id), message);
-}
-
-/**
- * Send a server query asking for the closing of a roll call
- * @param rollCallId the ID of the roll call
- * @param attendees list of the attendees' public keys
- * @param close (optional) the time at which the operation happens, defaults to now.
- * @return A Promise resolving when the operation is completed
- */
-export function requestCloseRollCall(
-  rollCallId: Hash,
-  attendees: PublicKey[],
-  close?: Timestamp,
-): Promise<void> {
-  const lao: Lao = OpenedLaoStore.get();
-  const time = (close === undefined) ? Timestamp.EpochNow() : close;
-
-  const message = new CloseRollCall({
-    update_id: Hash.fromStringArray(
-      EventTags.ROLL_CALL, lao.id.toString(), rollCallId.toString(), time.toString(),
-    ),
-    closes: rollCallId,
-    closed_at: time,
-    attendees: attendees,
-  });
-
-  return publish(channelFromIds(lao.id), message);
 }
