+++ fe1-web/store/reducers/__tests__/EventsReducer.test.ts
@@ -1,10 +1,9 @@
 import 'jest-extended';
 import { AnyAction } from 'redux';
-import {
-  Hash, LaoEventType, RollCall, RollCallState, RollCallStatus, Timestamp,
-} from 'model/objects';
+import { Hash, LaoEventType, Timestamp } from 'model/objects';
 import { mockLaoId } from '__tests__/utils/TestUtils';
 import { Meeting, MeetingState } from 'features/meeting/objects';
+import { RollCall, RollCallState, RollCallStatus } from 'features/rollCall/objects';
 import {
   addEvent,
   clearAllEvents,
